Add StationServices properties and events in line with 3.3.

+++ Howatworks.PlayerJournal.Serialization/Startup/Passengers.cs
@@ -11,7 +11,7 @@ namespace Howatworks.PlayerJournal.Serialization.Startup
         public class ManifestItem
         {
             [SuppressMessage("ReSharper", "InconsistentNaming")]
-            public int MissionID { get; set; } // TODO: check data type
+            public long MissionID { get; set; } // TODO: check data type
             public string Type { get; set; } // NOTE: type of passenger // TODO: check data type
             public bool VIP { get; set; }
             public bool Wanted { get; set; }
